Change fingerprinting approach
Controllers requiring a fingerprint can require `ensureFingerprint.php`.
If there is no client fingerprint in the session variable,
this script will redirect to `fingerprinter.php` to get a fingerprint,
before returning to the previous page.

This opens the door to refreshing the session, by removing the reliance
on the login controller to